+++ test/test.py
@@ -4,6 +4,7 @@ from game.model import (
     Tile,
     Square,
     Board,
+    Player,
 )
 from unittest.mock import patch
 
@@ -76,10 +77,18 @@ class TestSquare(unittest.TestCase):
     def test_adding_tile(self):
         square = Square(4, 5, multiplier_type='')
         letter = Tile(letter='X', value='10')
-        square.add_letter(letter=letter)
+        square.add_tile(letter=letter)
         self.assertEqual(square.tile, letter)
-
     
+    def test_none(self):
+        square = Square (4, 6, multiplier_type='')
+        score = square.calculate_score()
+        self.assertEqual(
+            score,
+            0,
+        )
+
+
     def test_multuplierLx2(self):
         square = Square(4, 5, multiplier_type='DL')
         square.tile = ('Z', 10)
@@ -100,7 +109,7 @@ class TestSquare(unittest.TestCase):
 class TestBoard(unittest.TestCase):
     def test_init_(self):
         board = Board ()
-        self. assertEqual(
+        self.assertEqual(
             len(board.grid),
             15,
         )
@@ -109,7 +118,20 @@ class TestBoard(unittest.TestCase):
             15,
         )
 
-        
+class TestPlayer(unittest.TestCase):
+    def test_init(self):
+        player = Player()
+        self.assertEqual(
+            len(player.playertiles),
+            0,
+        )
+    def test_starting_tiles(self):
+        player = Player()
+        tiles = player.starting_tiles()
+        self.assertEqual(
+            len(tiles),
+            7,
+        )
 
 if __name__ == '__main__':
     unittest.main()