+++ Changelog.md
@@ -35,3 +35,10 @@
 - New class added Board
 - Added CircleCi and env
 - New method in Square add_tile()
+
+## [0.0.5] 27-08-2023
+
+### Added
+
+- New class added Player
+- method starting_tiles()